[PCI_BIND] Verify correct linkage of device <> record for all device & record registry entries
- Add utility to verify correct linkage between the device and the record regisrty
- Add helper to get the device object from record
- Convert uint8 to uint32 for some variables
- Add test for verifying correct device <> record linkage

+++ kernel/arch/x86/pci/pci_driver_registry.c
@@ -32,7 +32,7 @@ uint8_t pci_driver_registry_add(pci_driver_registry_t *reg,
     }
 
     /* check if driver already has been registered */
-    for (uint8_t i = 0; i < reg->count; i++) {
+    for (uint32_t i = 0; i < reg->count; i++) {
         if (reg->drivers[i] == driver) {
             KLOG_ERROR(
                 "PCI_DRIVER",
